+++ CotEditor/Sources/Text Finder/Multiple Replace/Views/MultipleReplaceListView.swift
@@ -141,7 +141,7 @@ struct MultipleReplaceListView: View {
             Button(String(localized: "Action.replace.label", defaultValue: "Replace")) {
                 self.importingError = nil
                 do {
-                    try self.manager.importSetting(item.item, name: item.name, overwrite: true)
+                    try self.manager.importSetting(item.item, name: item.name, type: item.type, overwrite: true)
                 } catch {
                     self.error = error
                 }